VMM/IEM: Split the TLB into non-global (even) and global (odd) entries, doubling it in size.  In native code the global entries are only checked for ring-0 TBs, as checking both entries is slower than just the even one. bugref:10687
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@105036 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/include/IEMN8veRecompilerEmit.h
@@ -6898,7 +6898,7 @@ DECL_INLINE_THROW(void) iemNativeFixupFixedJump(PIEMRECOMPILERSTATE pReNative, u
     if ((uint8_t)(bOpcode - 0x70) < (uint8_t)0x10 || bOpcode == 0xeb)
     {
         pbCodeBuf[offFixup + 1] = (uint8_t)(offTarget - (offFixup + 2));
-        AssertStmt(pbCodeBuf[offFixup + 1] == offTarget - (offFixup + 2),
+        AssertStmt((int8_t)pbCodeBuf[offFixup + 1] == (int32_t)(offTarget - (offFixup + 2)),
                    IEMNATIVE_DO_LONGJMP(pReNative, VERR_IEM_EMIT_FIXED_JUMP_OUT_OF_RANGE));
     }
     else
